LayoutTests/accessibility/mac/progress-element-min-max.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<progress id="progress" value="50" max="75">Foo</progress>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the min/max values for a native progress element are exposed to the Mac platform.");

    if (window.accessibilityController) {

        var progress = accessibilityController.accessibleElementById("progress");
        shouldBe("progress.maxValue", "75");
        // There's actually no min attribute for <progress> elements, so this is always zero.
        shouldBe("progress.minValue", "0");
    }

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
