LayoutTests/accessibility/mac/progress-element-min-max-expected.txt


This tests that the min/max values for a native progress element are exposed to the Mac platform.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS progress.maxValue is 75
PASS progress.minValue is 0
PASS successfullyParsed is true

TEST COMPLETE
